Connection: move the connection owner deletion notice to the Connection package
A notice displays when a user attempts to delete the connection owner. Move this notice
and the endpoint for setting the connection owner to the connection package. With this change,
the connection owner deletion notice will be displayed for all consumers of the connection package.
