+++ packages/jitm/src/class-post-connection-jitm.php
@@ -38,30 +38,6 @@ class Post_Connection_JITM extends JITM {
 	}
 
 	/**
-	 * Prepare actions according to screen and post type.
-	 *
-	 * @since 3.8.2
-	 *
-	 * @uses Jetpack_Autoupdate::get_possible_failures()
-	 *
-	 * @param \WP_Screen $screen WP Core's screen object.
-	 */
-	public function prepare_jitms( $screen ) {
-		parent::prepare_jitms( $screen );
-		if ( ! in_array(
-			$screen->id,
-			array(
-				'jetpack_page_akismet-key-config',
-				'admin_page_jetpack_modules',
-			),
-			true
-		) ) {
-			// Not really a JITM. Don't know where else to put this :) .
-			add_action( 'admin_notices', array( $this, 'delete_user_update_connection_owner_notice' ) );
-		}
-	}
-
-	/**
 	 * A special filter for WooCommerce, to set a message based on local state.
 	 *
 	 * @param string $content The current message.
@@ -160,166 +136,12 @@ class Post_Connection_JITM extends JITM {
 	/**
 	 * This is an entire admin notice dedicated to messaging and handling of the case where a user is trying to delete
 	 * the connection owner.
+	 *
+	 * @deprecated since version 9.0.0. Use Automattic\Jetpack\Connection\Owner::delete_user_update_connection_owner_notice instead.
 	 */
 	public function delete_user_update_connection_owner_notice() {
-		global $current_screen;
-
-		/*
-		 * phpcs:disable WordPress.Security.NonceVerification.Recommended
-		 *
-		 * This function is firing within wp-admin and checks (below) if it is in the midst of a deletion on the users
-		 * page. Nonce will be already checked by WordPress, so we do not need to check ourselves.
-		 */
-
-		if ( ! isset( $current_screen->base ) || 'users' !== $current_screen->base ) {
-			return;
-		}
-
-		if ( ! isset( $_REQUEST['action'] ) || 'delete' !== $_REQUEST['action'] ) {
-			return;
-		}
-
-		// Get connection owner or bail.
-		$connection_manager  = new Manager();
-		$connection_owner_id = $connection_manager->get_connection_owner_id();
-		if ( ! $connection_owner_id ) {
-			return;
-		}
-		$connection_owner_userdata = get_userdata( $connection_owner_id );
-
-		// Bail if we're not trying to delete connection owner.
-		$user_ids_to_delete = array();
-		if ( isset( $_REQUEST['users'] ) ) {
-			$user_ids_to_delete = array_map( 'sanitize_text_field', wp_unslash( $_REQUEST['users'] ) );
-		} elseif ( isset( $_REQUEST['user'] ) ) {
-			$user_ids_to_delete[] = sanitize_text_field( wp_unslash( $_REQUEST['user'] ) );
-		}
-
-		// phpcs:enable
-		$user_ids_to_delete        = array_map( 'absint', $user_ids_to_delete );
-		$deleting_connection_owner = in_array( $connection_owner_id, (array) $user_ids_to_delete, true );
-		if ( ! $deleting_connection_owner ) {
-			return;
-		}
-
-		// Bail if they're trying to delete themselves to avoid confusion.
-		if ( get_current_user_id() === $connection_owner_id ) {
-			return;
-		}
-
-		// Track it!
-		if ( method_exists( $this->tracking, 'record_user_event' ) ) {
-			$this->tracking->record_user_event( 'delete_connection_owner_notice_view' );
-		}
-
-		$connected_admins = $connection_manager->get_connected_users( 'jetpack_disconnect' );
-		$user             = is_a( $connection_owner_userdata, 'WP_User' ) ? esc_html( $connection_owner_userdata->data->user_login ) : '';
-
-		echo "<div class='notice notice-warning' id='jetpack-notice-switch-connection-owner'>";
-		echo '<h2>' . esc_html__( 'Important notice about your Jetpack connection:', 'jetpack' ) . '</h2>';
-		echo '<p>' . sprintf(
-			/* translators: WordPress User, if available. */
-			esc_html__( 'Warning! You are about to delete the Jetpack connection owner (%s) for this site, which may cause some of your Jetpack features to stop working.', 'jetpack' ),
-			esc_html( $user )
-		) . '</p>';
-
-		if ( ! empty( $connected_admins ) && count( $connected_admins ) > 1 ) {
-			echo '<form id="jp-switch-connection-owner" action="" method="post">';
-			echo "<label for='owner'>" . esc_html__( 'You can choose to transfer connection ownership to one of these already-connected admins:', 'jetpack' ) . ' </label>';
-
-			$connected_admin_ids = array_map(
-				function( $connected_admin ) {
-						return $connected_admin->ID;
-				},
-				$connected_admins
-			);
-
-			wp_dropdown_users(
-				array(
-					'name'    => 'owner',
-					'include' => array_diff( $connected_admin_ids, array( $connection_owner_id ) ),
-					'show'    => 'display_name_with_login',
-				)
-			);
-
-			echo '<p>';
-			submit_button( esc_html__( 'Set new connection owner', 'jetpack' ), 'primary', 'jp-switch-connection-owner-submit', false );
-			echo '</p>';
-
-			echo "<div id='jp-switch-user-results'></div>";
-			echo '</form>';
-			?>
-			<script type="text/javascript">
-				jQuery( document ).ready( function( $ ) {
-					$( '#jp-switch-connection-owner' ).on( 'submit', function( e ) {
-						var formData = $( this ).serialize();
-						var submitBtn = document.getElementById( 'jp-switch-connection-owner-submit' );
-						var results = document.getElementById( 'jp-switch-user-results' );
-
-						submitBtn.disabled = true;
-
-						$.ajax( {
-							type        : "POST",
-							url         : "<?php echo esc_url( get_rest_url() . 'jetpack/v4/connection/owner' ); ?>",
-							data        : formData,
-							headers     : {
-								'X-WP-Nonce': "<?php echo esc_js( wp_create_nonce( 'wp_rest' ) ); ?>",
-							},
-							success: function() {
-								results.innerHTML = "<?php esc_html_e( 'Success!', 'jetpack' ); ?>";
-								setTimeout( function() {
-									$( '#jetpack-notice-switch-connection-owner' ).hide( 'slow' );
-								}, 1000 );
-							}
-						} ).done( function() {
-							submitBtn.disabled = false;
-						} );
-
-						e.preventDefault();
-						return false;
-					} );
-				} );
-			</script>
-			<?php
-		} else {
-			echo '<p>' . esc_html__( 'Every Jetpack site needs at least one connected admin for the features to work properly. Please connect to your WordPress.com account via the button below. Once you connect, you may refresh this page to see an option to change the connection owner.', 'jetpack' ) . '</p>';
-			$connect_url = \Jetpack::init()->build_connect_url( false, false, 'delete_connection_owner_notice' );
-			echo "<a href='" . esc_url( $connect_url ) . "' target='_blank' rel='noopener noreferrer' class='button-primary'>" . esc_html__( 'Connect to WordPress.com', 'jetpack' ) . '</a>';
-		}
-
-		echo '<p>';
-		printf(
-			wp_kses(
-				/* translators: URL to Jetpack support doc regarding the primary user. */
-				__( "<a href='%s' target='_blank' rel='noopener noreferrer'>Learn more</a> about the connection owner and what will break if you do not have one.", 'jetpack' ),
-				array(
-					'a' => array(
-						'href'   => true,
-						'target' => true,
-						'rel'    => true,
-					),
-				)
-			),
-			esc_url( Redirect::get_url( 'jetpack-support-primary-user' ) )
-		);
-		echo '</p>';
-		echo '<p>';
-		printf(
-			wp_kses(
-				/* translators: URL to contact Jetpack support. */
-				__( 'As always, feel free to <a href="%s" target="_blank" rel="noopener noreferrer">contact our support team</a> if you have any questions.', 'jetpack' ),
-				array(
-					'a' => array(
-						'href'   => true,
-						'target' => true,
-						'rel'    => true,
-					),
-				)
-			),
-			esc_url( Redirect::get_url( 'jetpack-contact-support' ) )
-		);
-		echo '</p>';
-		echo '</div>';
+		deprecated_function( __METHOD__, 'jetpack-9.0.0', 'Automattic\\Jetpack\\Connection\\Owner::delete_user_update_connection_owner_notice' );
+		return ( new Automattic\Jetpack\Connection\Owner() )->delete_user_update_connection_owner_notice();
 	}
 
 	/**
